apps/api/requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Database and storage
psycopg2-binary>=2.9,<3.0
django-storages>=1.14,<2.0
dj-database-url>=2.1,<3.0

# API and CORS
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
drf-spectacular>=0.27,<1.0

# Task queue
celery>=5.3,<6.0
django-celery-beat>=2.5,<3.0
redis>=5.0,<6.0

# AWS integration
boto3>=1.34,<2.0
django-ses>=3.5,<4.0

# SMS integration
twilio>=9.0,<10.0

# Image processing (basic)
Pillow>=10.1,<11.0

# Configuration and deployment
python-decouple>=3.8,<4.0
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0

# Billing and payments
stripe>=8.0,<14.0